Refactor LDG infrastructure to reduce circular dependencies
