+++ src/corecel/math/Quantity.hh
@@ -17,13 +17,6 @@
 namespace celeritas
 {
 //---------------------------------------------------------------------------//
-namespace detail
-{
-template<class, class>
-struct LdgLoader;
-}
-
-//---------------------------------------------------------------------------//
 /*!
  * A numerical value tagged with a unit.
  * \tparam UnitT  unit tag class
@@ -121,10 +114,11 @@ class Quantity
 #undef CELER_DEFINE_QACCESS
     //!@}
 
+    //! Access the underlying data for more efficient loading from memory
+    value_type const* data() const { return &value_; }
+
   private:
     value_type value_{};
-
-    friend detail::LdgLoader<Quantity<unit_type, value_type> const, void>;
 };
 
 //---------------------------------------------------------------------------//
@@ -328,24 +322,5 @@ inline char const* accessor_unit_label()
     return detail::AccessorResultType<T>::unit_type::label();
 }
 
-namespace detail
-{
-//! Template matching to determine if T is a Quantity
-template<class T>
-struct IsQuantity : std::false_type
-{
-};
-template<class V, class S>
-struct IsQuantity<Quantity<V, S>> : std::true_type
-{
-};
-template<class V, class S>
-struct IsQuantity<Quantity<V, S> const> : std::true_type
-{
-};
-template<class T>
-inline constexpr bool is_quantity_v = IsQuantity<T>::value;
-
 //---------------------------------------------------------------------------//
-}  // namespace detail
 }  // namespace celeritas
