+++ src/orange/univ/detail/LogicEvaluator.hh
@@ -9,6 +9,7 @@
 
 #include "corecel/Assert.hh"
 #include "corecel/cont/Span.hh"
+#include "corecel/data/LdgIterator.hh"
 #include "orange/OrangeTypes.hh"
 
 #include "LogicStack.hh"
